+++ demod/mod/demod_mod.h
@@ -1,5 +1,3 @@
-
-
 #include <math.h>
 #include <complex.h>
 
@@ -169,7 +167,7 @@ typedef struct {
 } hdb_t;
 
 
-float read_wav_header(pcm_t *, FILE *);
+int read_wav_header(pcm_t *, FILE *);
 int f32buf_sample(dsp_t *, int);
 int read_slbit(dsp_t *, int*, int, int, int, float, int);
 int read_softbit(dsp_t *, hsbit_t *, int, int, int, float, int);
@@ -183,5 +181,3 @@ int find_header(dsp_t *, float, int, int, int);
 int f32soft_read(FILE *fp, float *s);
 int find_binhead(FILE *fp, hdb_t *hdb, float *score);
 int find_softbinhead(FILE *fp, hdb_t *hdb, float *score);
-
-
